+++ dyndnsc/cli.py
@@ -72,37 +72,37 @@ def daemonize(stdout='/dev/null', stderr=None, stdin=os.devnull,
 
 
 def main():
-    from optparse import OptionParser
-    parser = OptionParser()
-    parser.add_option("-d", "--daemon", dest="daemon", help="go into daemon mode (implies --loop)", action="store_true", default=False)
-    parser.add_option("--hostname", dest="hostname", help="hostname to update", default=None)
-    parser.add_option("--key", dest="key", help="your authentication key", default=None)
-    parser.add_option("--userid", dest="userid", help="your userid", default=None)
-    parser.add_option("--password", dest="password", help="your password", default=None)
-    parser.add_option("--protocol", dest="protocol", help="protocol/service to use for updating your IP (default dyndns)", default='dyndns')
-    parser.add_option("--method", dest="method", help="method for detecting your IP (default webcheck)", default='webcheck')
-    parser.add_option("--loop", dest="loop", help="loop forever (default is to update once)", action="store_true", default=False)
-    parser.add_option("--sleeptime", dest="sleeptime", help="how long to sleep between checks in seconds", default=300)
-    parser.add_option("--version", dest="version", help="show version and exit", action="store_true", default=False)
-    (options, dummyargs) = parser.parse_args()
+    import argparse
+    parser = argparse.ArgumentParser()
+    parser.add_argument("-d", "--daemon", dest="daemon", help="go into daemon mode (implies --loop)", action="store_true", default=False)
+    parser.add_argument("--hostname", dest="hostname", help="hostname to update", default=None)
+    parser.add_argument("--key", dest="key", help="your authentication key", default=None)
+    parser.add_argument("--userid", dest="userid", help="your userid", default=None)
+    parser.add_argument("--password", dest="password", help="your password", default=None)
+    parser.add_argument("--protocol", dest="protocol", help="protocol/service to use for updating your IP (default dyndns)", default='dyndns')
+    parser.add_argument("--method", dest="method", help="method for detecting your IP (default webcheck)", default='webcheck')
+    parser.add_argument("--loop", dest="loop", help="loop forever (default is to update once)", action="store_true", default=False)
+    parser.add_argument("--sleeptime", dest="sleeptime", help="how long to sleep between checks in seconds", default=300)
+    parser.add_argument("--version", dest="version", help="show version and exit", action="store_true", default=False)
+    args = parser.parse_args()
 
-    if options.version:
+    if args.version:
         print("dyndnsc", pkg_resources.get_distribution("dyndnsc").version)  # pylint: disable=E1103
         return 0
 
     logging.basicConfig(level=logging.INFO, format='%(asctime)s %(levelname)s %(message)s')
 
-    if options.hostname is None:
+    if args.hostname is None:
         raise Exception("Please specify a hostname using --hostname")
 
     config = {}
-    config['hostname'] = options.hostname
-    config['key'] = options.key
-    config['userid'] = options.userid
-    config['password'] = options.password
-    config['protocol'] = options.protocol
-    config['method'] = options.method
-    config['sleeptime'] = int(options.sleeptime)
+    config['hostname'] = args.hostname
+    config['key'] = args.key
+    config['userid'] = args.userid
+    config['password'] = args.password
+    config['protocol'] = args.protocol
+    config['method'] = args.method
+    config['sleeptime'] = int(args.sleeptime)
 
     # done with command line options, bring on the dancing girls
     dyndnsclient = getDynDnsClientForConfig(config)
@@ -111,11 +111,11 @@ def main():
     # do an initial syncronisation, before going into endless loop:
     dyndnsclient.sync()
 
-    if options.daemon:
+    if args.daemon:
         daemonize()  # fork into background
-        options.loop = True
+        args.loop = True
 
-    if options.loop:
+    if args.loop:
         dyndnsclient.loop()
     else:
         dyndnsclient.check()